Consistent type for the parameters of GetStackTrace.

+++ src/stacktrace_win-inl.h
@@ -5,7 +5,7 @@
 _START_GOOGLE_NAMESPACE_
 
 // If you change this function, also change GetStackFrames below.
-size_t GetStackTrace(void** result, size_t max_depth, int skip_count) {
+int GetStackTrace(void** result, int max_depth, int skip_count) {
 // NOTE(phl): This was tuned for the static library with VS2013.
 #ifdef _DEBUG
   skip_count += 3;
